feat: Server-AI: Look handler function an unit tests

+++ Server/includes/ai/handlers/utils.h
@@ -9,9 +9,9 @@
 
 #include "world.h"
 
-player_t *get_player(linked_list_t players, int id);
+player_t *get_player(linked_list_t players, unsigned int id);
 
-int *next_case(world_t *world_t, int *coords, int orientation, int inc);
+size_t *next_case(world_t *world_t, size_t *coords, int orientation, int inc);
 
 int is_time_limit_passed(const clock_t start, const uint16_t limit_cycles,
         const unsigned int f);
